update_gitlab-ci.sh: add support for exceptions
In some cases, the diff between the defconfig we generated, and the
resulting defconfig from Buildroot has rejects. Very often, it means
that the configuration cannot be built.

However, in some cases, it makes sense to ignore some rejects. A
typical example is when building a bleeding edge toolchain for
ARC. When building a bleeding edge toolchain, we force the latest
version of gcc and binutils. However, ARC has only one version if
gcc/binutils, so those "bleeding edge" upstream versions are not
available, generating a reject. But we still want to be able to
generate bleeding edge ARC toolchains, that are based on a more recent
Buildroot version than "stable" ARC toolchains.

To achieve this, when there is a non-empty reject file,
update_gitlab-ci.sh will compare it with an exception file in
configs/exceptions/, and if they match (in terms of file name and
contents), then we will consider the configuration as valid.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>