Rework build scripts, first step
Instead of cloning Buildroot outside the chroot *and* inside the
chroot, we clone it only outside the chroot, and bind mount it inside
the chroot.

Also, chroot_build.sh is reworked with more functions, and the main
code at the bottom. Build now takes places from /opt instead of /tmp,
and generally the logic is cleaned up.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>