update_gitlab-ci.sh: Revert to one pipeline by trigged build
Commit [1] reworked update_gitlab-ci.sh to workaround a limit
on the number of jobs created for one pipeline. One pipeline
is created for each architecture.

In a follup patch, we will use manual trigged pipeline but
having to create each pipeline manually for each architecture
will not be very convenient.

[1] 8a6e2708ade17b11ca1f2c77048d5cecb3f4f00a

Signed-off-by: Romain Naour <romain.naour@gmail.com>