.gitlab-ci/generate-gitlab-ci-yml: rename variables provided by the parent pipeline to the child pipeline
TOOLCHAIN_BUILDER_* variables will be used as pre-filed variables in a
following patch. Pre-filled variables allows to the user to customize
a variable before running the pipeline.

For some reason (bug?) the value provided by the user is not provided
to a child pipeline and the default value is used instead.

So, rename TOOLCHAIN_BUILDER used in the child pipeline to avoid such
issue.

Signed-off-by: Romain Naour <romain.naour@gmail.com>