configs/extra/x86-64-v*: prevent building stable versions
The x86-64-{v2,v3,v4} can only be built with gcc 11.x, so for now we
don't want stable versions to be built. To prevent such stable
versions from being built, this extra file ensures that the stable
versions have GCC 10.x selected. Since it won't be the case, they will
be rejected, as we expect.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>