gitlab-ci: move the docker builder image to gitlab registry
In order to run a toolchain generated by toolchain-builder project on
most of all modern Linux distribution, we have to use the oldest
supported Linux distribution.

Initially, the toolchain-builder project built this old distribution
in a chroot for each build. Later, the chroot was replaced by a
Docker image placed in a temporary/personal dockerhub [1].

Since the commit [2] is merged, we can now move the docker image
to the toolchain-builder docker registry.

[1] https://hub.docker.com/r/kubu93/toolchain-builder
[2] ee063f1eefe41428afb33f4f9caedec77ba4c4d1

Signed-off-by: Romain Naour <romain.naour@gmail.com>